requirements.txt
asgiref==3.5.2
coverage==6.4.2
Django==4.0.6
django-environ==0.9.0
django-filter==22.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
gunicorn==20.1.0
PyJWT==2.4.0
PyMySQL==1.0.2
pytz==2022.1
sqlparse==0.4.2
tzdata==2022.1
